+++ plugin/src/main/kotlin/sc/plugin2023/Board.kt
@@ -16,38 +16,10 @@ class Board(fields: TwoDBoard<Field> = generateFields()): RectangularBoard<Field
     
     constructor(board: Board): this(board.gameField.clone())
     
-    /**
-     * Setzt einen Pinguin an gewählte Koordinaten. Diese Methode ist nur für
-     * den Server relevant, da keine vollständige Überprüfung auf korrekte Züge
-     * durchgeführt wird. Um für einen Spieler einen neuen Pinguin zu setzen,
-     * die [perform][sc.plugin2015.SetMove.perform]
-     * -Methode benutzen
-     *
-     * @param x
-     * X-Koordinate
-     * @param y
-     * Y-Koordinate
-     * @param penguin
-     * Pinguin, der gesetzt werden soll.
-     */
-    //@Throws(IllegalArgumentException::class)
-    //fun putPenguin(x: Int, y: Int, penguin: Penguin?) {
-    //    require((x < 0 || y < 0 || x >= Constants.BOARD_SIZE || y >= Constants.BOARD_SIZE || fields[x][y].fish !== 1 || fields[x][y].penguin) == null)
-    //    fields[x][y].putPenguin(penguin)
-    //}
-    ///**
-    // * nur für den Server relevant
-    // */
-    //@Throws(IllegalArgumentException::class)
-    //private fun putPenguinMove(x: Int, y: Int, penguin: Penguin) {
-    //    require((x < 0 || y < 0 || x >= Constants.BOARD_SIZE || y >= Constants.BOARD_SIZE || fields[x][y].fish <= 0 || fields[x][y].penguin) == null)
-    //    fields[x][y].putPenguin(penguin)
-    //}
-
     override fun isValid(coordinates: Coordinates) =
             (coordinates.x + coordinates.y) % 2 == 0 &&
-                coordinates.x >= 0 &&
-                super.isValid(coordinates.copy(coordinates.x / 2))
+            coordinates.x >= 0 &&
+            super.isValid(coordinates.copy(coordinates.x / 2))
     
     /** Gibt das Feld an den gegebenen Koordinaten zurück. */
     override operator fun get(x: Int, y: Int) =
@@ -64,11 +36,11 @@ class Board(fields: TwoDBoard<Field> = generateFields()): RectangularBoard<Field
     }
     
     fun possibleMovesFrom(pos: Coordinates) =
-        Vector.DoubledHex.directions.flatMap { vector ->
-            (1 until Constants.BOARD_SIZE).map {
-                Move.run(pos, vector * it)
-            }.takeWhile { getOrEmpty(it.to).fish > 0 }
-        }
+            Vector.DoubledHex.directions.flatMap { vector ->
+                (1 until Constants.BOARD_SIZE).map {
+                    Move.run(pos, vector * it)
+                }.takeWhile { getOrEmpty(it.to).fish > 0 }
+            }
     
     /** Returns a list of the non-null filter outputs */
     fun <T> filterFields(filter: (Field, Coordinates) -> T?): Collection<T> =
@@ -86,21 +58,28 @@ class Board(fields: TwoDBoard<Field> = generateFields()): RectangularBoard<Field
     fun getOrEmpty(key: Coordinates?) = key?.let { getOrNull(it) } ?: Field()
     
     override val entries: Set<Map.Entry<Coordinates, Field>>
-        get() = filterFields { field, coordinates -> FieldPosition(coordinates, field) }.toSet()
+        get() = filterFields { f, coordinates -> FieldPosition(coordinates, f) }.toSet()
     
     override fun clone(): Board = Board(this)
     
     companion object {
         /** Generiert ein neues Spielfeld mit zufällig auf dem Spielbrett verteilten Fischen. */
         private fun generateFields(seed: Int = Random.nextInt()): TwoDBoard<Field> {
-            var remainingFish = 100
+            var remainingFish = Constants.BOARD_SIZE * (Constants.BOARD_SIZE + 1)
             val random = Random(seed)
-            // TODO val holes =
-            return List(Constants.BOARD_SIZE) {
+            println("Board seed: $seed")
+            return List(Constants.BOARD_SIZE / 2) {
                 MutableList(Constants.BOARD_SIZE) {
-                    val fish = random.nextInt(remainingFish) / 30 + 1
+                    val rand = random.nextInt(remainingFish)
+                    if(rand < 5)
+                        return@MutableList Field()
+                    val fish = rand / 20 + 1
                     remainingFish -= fish
                     Field(fish)
+                }
+            }.let {
+                it + it.asReversed().map { list ->
+                    MutableList(Constants.BOARD_SIZE) { index -> list[Constants.BOARD_SIZE - index - 1].clone() }
                 }
             }
         }